Modeler: refactor actions that use DataSourceWizard
